+++ src/bikes/core/schemas.py
@@ -51,7 +51,7 @@ class Schema(pa.DataFrameModel):
 class InputsSchema(Schema):
     """Schema for the project inputs."""
 
-    instant: papd.Index[padt.UInt32] = pa.Field(ge=0, check_name=True)
+    instant: papd.Index[padt.UInt32] = pa.Field(ge=0)
     dteday: papd.Series[padt.DateTime] = pa.Field()
     season: papd.Series[padt.UInt8] = pa.Field(isin=[1, 2, 3, 4])
     yr: papd.Series[padt.UInt8] = pa.Field(ge=0, le=1)
@@ -75,7 +75,7 @@ Inputs = papd.DataFrame[InputsSchema]
 class TargetsSchema(Schema):
     """Schema for the project target."""
 
-    instant: papd.Index[padt.UInt32] = pa.Field(ge=0, check_name=True)
+    instant: papd.Index[padt.UInt32] = pa.Field(ge=0)
     cnt: papd.Series[padt.UInt32] = pa.Field(ge=0)
 
 
@@ -85,8 +85,36 @@ Targets = papd.DataFrame[TargetsSchema]
 class OutputsSchema(Schema):
     """Schema for the project output."""
 
-    instant: papd.Index[padt.UInt32] = pa.Field(ge=0, check_name=True)
+    instant: papd.Index[padt.UInt32] = pa.Field(ge=0)
     prediction: papd.Series[padt.UInt32] = pa.Field(ge=0)
 
 
 Outputs = papd.DataFrame[OutputsSchema]
+
+
+class SHAPValuesSchema(Schema):
+    """Schema for the project shap values."""
+
+    class Config:
+        """Default configurations this schema.
+
+        Parameters:
+            dtype (str): dataframe default data type.
+            strict (bool): ensure the data type is correct.
+        """
+
+        dtype: str = "float32"
+        strict: bool = False
+
+
+SHAPValues = papd.DataFrame[SHAPValuesSchema]
+
+
+class FeatureImportancesSchema(Schema):
+    """Schema for the project feature importances."""
+
+    feature: papd.Series[padt.String] = pa.Field()
+    importance: papd.Series[padt.Float32] = pa.Field()
+
+
+FeatureImportances = papd.DataFrame[FeatureImportancesSchema]