+++ src/bikes/io/registries.py
@@ -34,17 +34,33 @@ def uri_for_model_alias(name: str, alias: str) -> str:
     return f"models:/{name}@{alias}"
 
 
-def uri_for_model_version(name: str, version: str) -> str:
+def uri_for_model_version(name: str, version: int) -> str:
     """Create a model URI from a model name and a version.
 
     Args:
         name (str): name of the mlflow registered model.
-        version (str): version of the registered model.
+        version (int): version of the registered model.
 
     Returns:
         str: model URI as "models:/name/version."
     """
     return f"models:/{name}/{version}"
+
+
+def uri_for_model_alias_or_version(name: str, alias_or_version: str | int) -> str:
+    """Create a model URi from a model name and an alias or version.
+
+    Args:
+        name (str): name of the mlflow registered model.
+        alias_or_version (str | int): alias or version of the registered model.
+
+    Returns:
+        str: model URI as "models:/name@alias" or "models:/name/version" based on input.
+    """
+    if isinstance(alias_or_version, int):
+        return uri_for_model_version(name=name, version=alias_or_version)
+    else:
+        return uri_for_model_alias(name=name, alias=alias_or_version)
 
 
 # %% SAVERS
@@ -113,7 +129,7 @@ class CustomSaver(Saver):
             Args:
                 context (mlflow.pyfunc.PythonModelContext): mlflow context.
                 model_input (schemas.Inputs): inputs for the mlflow model.
-                params(dict[str, T.Any] | None): additional parameters.
+                params (dict[str, T.Any] | None): additional parameters.
 
             Returns:
                 schemas.Outputs: validated outputs of the project model.