+++ src/bikes/jobs/__init__.py
@@ -2,6 +2,8 @@
 
 # %% IMPORTS
 
+from bikes.jobs.evaluations import EvaluationsJob
+from bikes.jobs.explanations import ExplanationsJob
 from bikes.jobs.inference import InferenceJob
 from bikes.jobs.promotion import PromotionJob
 from bikes.jobs.training import TrainingJob
@@ -9,8 +11,16 @@ from bikes.jobs.tuning import TuningJob
 
 # %% TYPES
 
-JobKind = TuningJob | TrainingJob | PromotionJob | InferenceJob
+JobKind = TuningJob | TrainingJob | PromotionJob | InferenceJob | EvaluationsJob | ExplanationsJob
 
 # %% EXPORTS
 
-__all__ = ["TuningJob", "TrainingJob", "PromotionJob", "InferenceJob", "JobKind"]
+__all__ = [
+    "TuningJob",
+    "TrainingJob",
+    "PromotionJob",
+    "InferenceJob",
+    "EvaluationsJob",
+    "ExplanationsJob",
+    "JobKind",
+]