+++ src/bikes/jobs/training.py
@@ -4,6 +4,7 @@
 
 import typing as T
 
+import mlflow
 import pydantic as pdt
 
 from bikes.core import metrics as metrics_
@@ -61,8 +62,9 @@ class TrainingJob(base.Job):
         logger.info("With logger: {}", logger)
         # - mlflow
         client = self.mlflow_service.client()
+        logger.info("With client: {}", client.tracking_uri)
         with self.mlflow_service.run_context(run_config=self.run_config) as run:
-            logger.info("With mlflow run id: {}", run.info.run_id)
+            logger.info("With run context: {}", run.info)
             # data
             # - inputs
             logger.info("Read inputs: {}", self.inputs)
@@ -74,6 +76,19 @@ class TrainingJob(base.Job):
             targets_ = self.targets.read()  # unchecked!
             targets = schemas.TargetsSchema.check(targets_)
             logger.debug("- Targets shape: {}", targets.shape)
+            # lineage
+            # - inputs
+            logger.info("Log lineage: inputs")
+            inputs_lineage = self.inputs.lineage(data=inputs, name="inputs")
+            mlflow.log_input(dataset=inputs_lineage, context=self.run_config.name)
+            logger.debug("- Inputs lineage: {}", inputs_lineage.to_dict())
+            # - targets
+            logger.info("Log lineage: targets")
+            targets_lineage = self.targets.lineage(
+                data=targets, name="targets", targets=schemas.TargetsSchema.cnt
+            )
+            mlflow.log_input(dataset=targets_lineage, context=self.run_config.name)
+            logger.debug("- Targets lineage: {}", targets_lineage.to_dict())
             # splitter
             logger.info("With splitter: {}", self.splitter)
             # - index
@@ -117,4 +132,8 @@ class TrainingJob(base.Job):
                 name=self.mlflow_service.registry_name, model_uri=model_info.model_uri
             )
             logger.debug("- Model version: {}", model_version)
+            # notify
+            self.alerts_service.notify(
+                title="Training Job Finished", message=f"Model version: {model_version.version}"
+            )
         return locals()