+++ src/bikes/jobs/tuning.py
@@ -4,6 +4,7 @@
 
 import typing as T
 
+import mlflow
 import pydantic as pdt
 
 from bikes.core import metrics, models, schemas
@@ -59,9 +60,8 @@ class TuningJob(base.Job):
         # - logger
         logger = self.logger_service.logger()
         logger.info("With logger: {}", logger)
-        # - mlflow
         with self.mlflow_service.run_context(run_config=self.run_config) as run:
-            logger.info("With mlflow run id: {}", run.info.run_id)
+            logger.info("With run context: {}", run.info)
             # data
             # - inputs
             logger.info("Read inputs: {}", self.inputs)
@@ -73,6 +73,19 @@ class TuningJob(base.Job):
             targets_ = self.targets.read()  # unchecked!
             targets = schemas.TargetsSchema.check(targets_)
             logger.debug("- Targets shape: {}", targets.shape)
+            # lineage
+            # - inputs
+            logger.info("Log lineage: inputs")
+            inputs_lineage = self.inputs.lineage(data=inputs, name="inputs")
+            mlflow.log_input(dataset=inputs_lineage, context=self.run_config.name)
+            logger.debug("- Inputs lineage: {}", inputs_lineage.to_dict())
+            # - targets
+            logger.info("Log lineage: targets")
+            targets_lineage = self.targets.lineage(
+                data=targets, name="targets", targets=schemas.TargetsSchema.cnt
+            )
+            mlflow.log_input(dataset=targets_lineage, context=self.run_config.name)
+            logger.debug("- Targets lineage: {}", targets_lineage.to_dict())
             # model
             logger.info("With model: {}", self.model)
             # metric
@@ -91,4 +104,8 @@ class TuningJob(base.Job):
             logger.debug("- Results: {}", results.shape)
             logger.debug("- Best Score: {}", best_score)
             logger.debug("- Best Params: {}", best_params)
+            # notify
+            self.alerts_service.notify(
+                title="Tuning Job Finished", message=f"Best score: {best_score}"
+            )
         return locals()