+++ tests/core/test_schemas.py
@@ -1,6 +1,6 @@
 # %% IMPORTS
 
-from bikes.core import schemas
+from bikes.core import models, schemas
 from bikes.io import datasets
 
 # %% SCHEMAS
@@ -31,3 +31,25 @@ def test_outputs_schema(outputs_reader: datasets.Reader) -> None:
     data = outputs_reader.read()
     # then
     assert schema.check(data) is not None, "Outputs data should be valid!"
+
+
+def test_shap_values_schema(
+    model: models.Model,
+    train_test_sets: tuple[schemas.Inputs, schemas.Targets, schemas.Inputs, schemas.Targets],
+) -> None:
+    # given
+    schema = schemas.SHAPValuesSchema
+    _, _, inputs_test, _ = train_test_sets
+    # when
+    data = model.explain_samples(inputs=inputs_test)
+    # then
+    assert schema.check(data) is not None, "SHAP values data should be valid!"
+
+
+def test_feature_importances_schema(model: models.Model) -> None:
+    # given
+    schema = schemas.FeatureImportancesSchema
+    # when
+    data = model.explain_model()
+    # then
+    assert schema.check(data) is not None, "Feature importance data should be valid!"