+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+## v1.1.0 (2024-07-21)
+
+### Feat
+
+- **kpi**: add key performance indicators
+- **mlproject**: add mlflow project and tasks
+- **monitoring**: add mlflow.evaluate API
+- **lineage**: add lineage features through mlflow data api
+- **explanations**: add explainability features and tooling
+- **data**: add train, test, and sample data
+- **notification**: add service and alerts with plyer
+- **observability**: add alerting with plyer notifications
+- **observability**: add infrastructure through mlflow system metrics
+
+### Fix
+
+- **kpi**: add key performance indicators
+- **projects**: change naming convention
+- **evaluation**: add evaluation files
+- **loading**: use version or alias for loading models
+- **warnings**: improve styles and remove warnings
+- **mlflow**: remove input examples following the addition of lineage
+- **paths**: fix path for explanation job
+- **data**: fix models explanations name
+- **data**: add parquet data
+
 ## v1.0.1 (2024-06-28)
 
 ### Fix